feat: React公式チュートリアルのリファクタリングが完了 #54
