+++ test_basics/11_Storybook/src/components/Game/Game.tsx
@@ -1,44 +1,22 @@
-import React, { useState } from 'react';
-import { calculateWinner } from '../../helpers';
-import { SquareValue } from '../../types';
+import React from 'react';
+import { BoardValue } from '../../types/tictactoe';
 import { Board } from '../Board/Board';
 
-export const Game: React.FC = () => {
-  const [xIsNext, setXIsNext] = useState<boolean>(true);
-  const [stepNumber, setStepNumber] = useState<number>(0);
-  const [history, setHistory] = useState<{ squares: SquareValue[] }[]>([
-    {
-      squares: Array(9).fill(null),
-    },
-  ]);
+export interface Props {
+  status: string;
+  current: BoardValue;
+  history: BoardValue[];
+  handleClick: (i: number) => void;
+  jumpTo: (step: number) => void;
+}
 
-  const handleClick = (i: number): void => {
-    const newHistory = history.slice(0, stepNumber + 1);
-    const current = newHistory[newHistory.length - 1];
-    const squares = current.squares.slice();
-    if (calculateWinner(squares) || squares[i]) {
-      return;
-    }
-    squares[i] = xIsNext ? 'X' : 'O';
-    setHistory(
-      newHistory.concat([
-        {
-          squares: squares,
-        },
-      ]),
-    );
-    setStepNumber(newHistory.length);
-    setXIsNext(!xIsNext);
-  };
-
-  const jumpTo = (step: number): void => {
-    setStepNumber(step);
-    setXIsNext(step % 2 === 0);
-  };
-
-  const current = history[stepNumber];
-  const winner = calculateWinner(current.squares);
-
+export const Game = ({
+  status,
+  current,
+  history,
+  handleClick,
+  jumpTo,
+}: Props): JSX.Element => {
   const moves = history.map((step, move) => {
     const desc = move ? 'Go to move #' + move : 'Go to game start';
     return (
@@ -47,13 +25,6 @@ export const Game: React.FC = () => {
       </li>
     );
   });
-
-  let status;
-  if (winner) {
-    status = 'Winner: ' + winner;
-  } else {
-    status = 'Next Player: ' + (xIsNext ? 'X' : 'O');
-  }
 
   return (
     <div className="game">
