--- test_basics/11_Storybook/src/types.tsx
@@ -1,1 +1,0 @@
-export type SquareValue = 'X' | 'O' | null;
