+++ test_basics/11_Storybook/src/types/tictactoe.ts
@@ -1,0 +1,5 @@
+export type SquareValue = 'X' | 'O' | null;
+
+export interface BoardValue {
+  squares: SquareValue[];
+}
